Created several actions in the games controller and some logic that allows new games to be created for users that aren't logged in and saved in their cookies so they might resume if they navigate away.
Also created an update action that accepts json and stores it in the game
database as text, and found out how to fetch it back out.

Still to be done: create the JS to create and load gamestates in JSON format.

I imagine much of the actual UI logic will have to be re-written for this to work,
I'm not looking forward to that.