Well I definitely rewrote more than just the home action. I created a new home action, some new sessions helper methods and a number of before actions on both the users and games controller.
The end result is that now users should be able to visit the homepage,
and if they're not logged in it will create a new puzzle for them.
If, after that they log in it will be bound to their User.
If they are logged in when they visit the root they'll be redirected to
their last created game.

The helper methods and before_action methods I created are to ensure that
the permissions are set right, and that users cannot play or update each other's games.

It seemed fitting to update the puzzles and users controllers once I had created that
permissions systems because they were outdated without it.